actors update
- transform component (position, rotation, scale)
- component selfloading
- actor properties reduction
- components factory
- get info components func

+++ Core/Physics/PhysicsManager.h
@@ -10,7 +10,7 @@ namespace liman
 		PhysicsManager() {}
 		virtual ~PhysicsManager() {}
 
-		void UpdateMovables(float deltaTime);
+		void UpdateMovables(int deltaTime);
 
 	private:
